# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-PERSON ",
        "description": "Documentação da API-PERSON com Swagger Service.",
        "contact": {
            "name": "Luiz Santos Full Stack Developer ",
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/servidores-efetivos": {
            "get": {
                "tags": [
                    "Servidor Efetivo"
                ],
                "summary": "Lista todos as servidores efetivos",
                "description": "Retorna uma lista de servidores efetivos armazenados no banco de dados.",
                "operationId": "52e4ce496fb4d158b46016deec180806",
                "responses": {
                    "200": {
                        "description": "Lista de servidores efetivos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ServidorEfetivoFull"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum servidor efetivo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Não há servidores efetivos cadastrados!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao buscar os servidores efetivos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar os servidores efetivos."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servidor Efetivo"
                ],
                "summary": "Cria um servidor efetivo",
                "description": "Registra um servidor efetivo no banco de dados.",
                "operationId": "db3aca000330b5d6d5334b33a2b8f38c",
                "requestBody": {
                    "description": "Dados necessários para criar um servidor efetivo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pes_nome",
                                    "pes_cpf",
                                    "se_matricula",
                                    "pes_data_nascimento",
                                    "pes_sexo",
                                    "pes_mae",
                                    "pes_pai",
                                    "cid_id",
                                    "end_tipo_logradouro",
                                    "end_logradouro",
                                    "end_numero",
                                    "end_bairro"
                                ],
                                "properties": {
                                    "pes_nome": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "pes_cpf": {
                                        "type": "string",
                                        "example": "111.222.333-44"
                                    },
                                    "se_matricula": {
                                        "type": "string",
                                        "example": "2003456788467"
                                    },
                                    "pes_data_nascimento": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "1978-08-23T12:00:00Z"
                                    },
                                    "pes_sexo": {
                                        "type": "string",
                                        "example": "Masculino"
                                    },
                                    "pes_mae": {
                                        "type": "string",
                                        "example": "Maria Aparecida da Silva"
                                    },
                                    "pes_pai": {
                                        "type": "string",
                                        "example": "Cícero Joaquim da Silva"
                                    },
                                    "cid_id": {
                                        "type": "integer",
                                        "example": "21"
                                    },
                                    "end_tipo_logradouro": {
                                        "type": "string",
                                        "example": "Avenida"
                                    },
                                    "end_logradouro": {
                                        "type": "string",
                                        "example": "Silviano Brandão"
                                    },
                                    "end_numero": {
                                        "type": "integer",
                                        "example": "1000"
                                    },
                                    "end_complemento": {
                                        "type": "string",
                                        "example": "Bloco E, 50 apartamento 303"
                                    },
                                    "end_bairro": {
                                        "type": "string",
                                        "example": "Horto Florestal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pessoa criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pessoa criado com sucesso."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Pessoa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao criar o servidor efetivo.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar o servidor efetivo."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pessoa": {
                "title": "Pessoa",
                "description": "Modelo de Pessoa",
                "required": [
                    "pes_nome",
                    "pes_cpf",
                    "pes_data_nascimento",
                    "pes_sexo",
                    "pes_mae",
                    "pes_pai"
                ],
                "properties": {
                    "pes_id": {
                        "description": "ID do Pessoa",
                        "type": "integer",
                        "example": 1
                    },
                    "pes_nome": {
                        "description": "Nome da pessoa",
                        "type": "string",
                        "example": "João da Silva"
                    },
                    "pes_cpf": {
                        "description": "CPF da pessoa",
                        "type": "string",
                        "example": "111.222.333-44"
                    },
                    "pes_data_nascimento": {
                        "description": "Data de nascimento",
                        "type": "string",
                        "format": "date-time",
                        "example": "1978-08-23"
                    },
                    "pes_sexo": {
                        "description": "Sexo da pessoa",
                        "type": "string",
                        "example": "Masculino"
                    },
                    "pes_mae": {
                        "description": "Nome da mãe",
                        "type": "string",
                        "example": "Maria Aparecida da Silva"
                    },
                    "pes_pai": {
                        "description": "Nome da pai",
                        "type": "string",
                        "example": "Cícero Joaquim da Silva"
                    }
                },
                "type": "object"
            },
            "ServidorEfetivoFull": {
                "properties": {
                    "pessoa": {
                        "$ref": "#/components/schemas/PessoaFull"
                    },
                    "criado_em": {
                        "type": "string",
                        "example": "24/03/2025 21:21"
                    }
                },
                "type": "object"
            },
            "PessoaFull": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 22
                    },
                    "nome": {
                        "type": "string",
                        "example": "João da Silva"
                    },
                    "matricula": {
                        "type": "string",
                        "example": "2003456788467"
                    },
                    "Data de nascimento": {
                        "type": "string",
                        "format": "date",
                        "example": "1978-08-23"
                    },
                    "Cadastro de Pessoa Físsica": {
                        "type": "string",
                        "example": "111.222.333-44"
                    },
                    "enderecos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EnderecoFull"
                        }
                    }
                },
                "type": "object"
            },
            "EnderecoFull": {
                "properties": {
                    "logradouro": {
                        "type": "string",
                        "example": "Silviano Brandão"
                    },
                    "numero": {
                        "type": "string",
                        "example": "1000"
                    },
                    "complemento": {
                        "type": "string",
                        "example": "Bloco E, 50 apartamento 303"
                    },
                    "bairro": {
                        "type": "string",
                        "example": "Horto Florestal"
                    },
                    "cidade_id": {
                        "type": "integer",
                        "example": 21
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Digite \"Bearer {token}\" para se autenticar.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}